personal-agent/requirements.txt
# Week 1 — Foundation
anthropic>=0.40.0
slack-bolt>=1.21.0
slack-sdk>=3.33.0
python-dotenv>=1.0.1
structlog>=24.4.0
rich>=13.9.0

# Week 2 — Intelligence
apscheduler>=3.10.4
tiktoken>=0.8.0
aiofiles>=24.1.0
aiosqlite>=0.20.0
pydantic>=2.10.0

# Week 3 — Integrations
playwright>=1.49.0
httpx>=0.28.0
duckduckgo-search>=7.2.0
beautifulsoup4>=4.12.3
lxml>=5.3.0

# Week 4 — Polish
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
mypy>=1.13.0
ruff>=0.8.0
watchdog>=6.0.0
